fixup! Added MOSTheme provider, moved types
