+++ config/storybook/decorators/themeDecorator.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import MosTheme from '../../../src/themes/MosTheme/MosTheme';
+import MosTheme from '../../../src/themes/MosTheme/MosTheme.template';
 
 const themeDecorator = (story:Function):JSX.Element => (
     <MosTheme>{story()}</MosTheme>
