+++ src/themes/MosTheme/index.ts
@@ -1,1 +1,1 @@
-export { default as default } from './MosTheme';
+export { default as default } from './MosTheme.template';
